Test project to demonstrate problem with setIsAccessibilityElement
Run on iOS 6.0 or 6.1 simulator: if accessibility inspector is off then
four 'NO' values are output to the log, if inspector is on then two
'YES', two 'NO are output -> see code in ViewController.m/viewDidLoad
for why that is wrong.

On a 6.1 device (iPhone or iPad) if the Home Button Triple Click ->
Voice Over is enabled then the same problem shows up, or if voice over
is on by another method. Otherwise it is fine.

Removing MessageUI framework from the app stops the problem - and that
framework is just linked in, not used.